Changed Token to abstract class and implemented OperatorToken and OperandToken.
TokenString is still a collection of tokens, but when parsing the
operator and operand tokens can now be treated differently.
TokenType was deleted as it was no longer needed.
Implemented TokeniserTest.
Implemented an Indexer for accessing Tokens from TokenString.
Fixed the next() and previous() methods from TokenString so they return
the Token at the current index, then increment.