Users can now define their own constants in config.json!
--changes--
Cleaned up Tokeniser code.
Added the ConstantToken (subclass of OperandToken).
Updated config.json and config.cs to read user-defined constants.

--comments--
This was much easier to implement than expected.
The code for parsing constants was very similar to the code for parsing
multi-character numbers.
Furthermore no change at all was made to the equation solving algorithm,
since the ConstantTokens could be cast to an OperandToken.

Of everything I've done so far, this is the feature I'm most proud of.
Thanks for reading :)