Updated operator and operation bindings in config.json
--changes--
config.json updated to bind operator and operations as an array of
arrays, rather than two arrays.
config.cs has been updated to reflect this.
removed unused unary operator binding.

--comments--
The new binding reflects the way user constants are defined, this not
only maintains consistency but also looks much cleaner.
It also removed the need to check if the number of operators is the same
as the number of operations.