Updated Tokens to have a position field.
--changes--
Token and its subclasses have new constructors that can take a position
of the characters within the equation.
Tokeniser has been updated to allow this.

--notes--
The position of single character tokens are just their position in the
equation (index+1). Multi-character values are assigned the position of
their first character, for example 567 would have the position of 5, and
-42 would have the position of the negation.

I was debating using position in the equation or position of the token,
but realised position of the token could simply be found with the
TokenString, so there is a lot of potential for feedback to users.

The non-position constructors are still there, but I will leave them
alone as someone working on this in the future may want them for some
reason. It would be annoying to try to create a token and have to give
it bogus positions if you didn't have one.

This is part of my attempt to implement equation validation along with
actionable feedback to users. The position of the token in the equation
could be used to inform a user where the error was.