Equations can now be validated before solving
--changes--
Implemented EqValidator. It can take an equation and validate it,
including identifying and reporting where the errors are.
EqValidator returns a ValidationInstance, which contains a flag for the
valdity of the equation and (if valid) a number of ValidationObjects
which represent validation failures (include a message and the offending
token)
Restructured validation code. Seperated the classes into their own files
under the EquationSolver.Validation namespace.
Added ValidationTests.cs

--comments--
It correctly identified when the number of opening and closing brackets
is not the same, but finding the offending bracket is not yet
implemented.
Validity of an operand is identified by checking if it is adjacent to
another operand - it should not be.
Validity of an operator is identified by checking if it has adjacent
operands, including through brackets.