EqValidator now gives the location of invalid brackets
--changes--
Finished EqValidators isvalid method.
Added a method to TokenString to set its index to the last value, this
is used for iterating right-left (used for finding unopened brackets.
Added method to validationInstance to order its collection of
ValidationObjects.

--comments--
I realised that because of the brackets being found last (due to it
being a costly search that would be inefficient do to in-time with the
rest of the identification) order would no longer be guaranteed, as an
unclosed bracket in position 1 would come after all non-brackets, and
because it uses a stack to find them the order of the invalid brackets
is also reversed.
To resolve these issues I used LinQ to order the collection from first
token to last, this method is called just before the return.