Completed the EqSolve API and integrated validation.
--changes--
The solving API has three paths:
 - The old path, solves the equation without validation and returns a
   decimal (can throw unfriendly exceptions)
 - Solves equation with validation and returns a decimal (throws a
   Validation Exception of validation fails)
 - solves Equation with validation and returns a SolveResult object

Added SolveResult. SolveResult contains a flag for the success of
solving an equation, the result (if successful) or the
ValidationInstance (if unsuccessful)

Added ValidationException. A subclass of Exception, is exactly the same
except it contains a field for the ValidationInstance

--comments--
This is likely the last big content update, there are other small
changes I may still like (such as covering additional scenarios in
validation) but it will mostly be bugfixes and small features from now
on.

I added a bunch of examples of how to use the API to Program.cs, as well
as some comments explaining them.

I amended the unit tests to work with the new API, but this is temporary
as I will be looking into migrating my tests to specflow.

Finally I also added some class summaries where I felt important, I'll
probably be adding more later on.